Update version to 1.2.0, refactor build scripts, and enhance server functionality. Added new Keycloak configuration and service layers, updated README with new commands, and removed deprecated index.ts file.
