+++ smithery.yaml
@@ -5,7 +5,7 @@ startCommand:
   commandFunction:
     # A JS function that produces the CLI command based on the given config to start the MCP on stdio.
     |-
-    (config) => ({ command: 'node', args: ['dist/index.js'], env: { KEYCLOAK_URL: config.keycloakUrl, KEYCLOAK_ADMIN: config.keycloakAdmin, KEYCLOAK_ADMIN_PASSWORD: config.keycloakAdminPassword } })
+    (config) => ({ command: 'node', args: ['dist/server.js'], env: { KEYCLOAK_URL: config.keycloakUrl, KEYCLOAK_ADMIN: config.keycloakAdmin, KEYCLOAK_ADMIN_PASSWORD: config.keycloakAdminPassword } })
   configSchema:
     # JSON Schema defining the configuration options for the MCP.
     type: object
